split out drifter
